Update README
Added more information on what to do with a pre-trained neural net.

I was battling with some gunzip errors until I realised that the gz file is expected as the -model argument and not the zip archive that is offered on the releases page. You need to unzip that before use.